created UI folder and moved ListGenerationUI file
